pyproject.toml
[project]
name = "ai-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.12.0",
    "ipykernel>=7.1.0",
    "langchain==0.3.2",
    "langchain-community>=0.3.1",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "wikipedia>=1.4.0",
]
